requirements.txt
numpy==2.3.5
pandas==2.3.3
Pillow==12.0.0
pycountry==24.6.1
pydeck==0.9.1
streamlit==1.51.0
